created data dir, moved data objects and json to folder.  Updated relative imports
