+++ src/Styles/Challenges.css
@@ -85,14 +85,23 @@
   width: 100%;
   height: 100%;
   border-radius: 1rem;
-  padding: 3rem;
   backface-visibility: hidden;
   outline: 2px solid rgb(71, 71, 71);
 }
 
 .front {
+  padding: 1.5rem;
   background: whitesmoke;
   color: black;
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.front > img {
+  width: 100%;
+  height: 150px;
+  background-color: grey;
 }
 
 .back {
@@ -113,7 +122,7 @@
 }
 
 @media screen and (max-width: 500px) {
-  .card-sleeve {
+  .card-placeholder {
     width: 250px;
     height: 375px;
   }
